@@ -46,7 +46,7 @@
     "electron-builder": "21.2.0",
     "electron-debug": "1.5.0",
     "electron-devtools-installer": "2.2.3",
-    "electron-log": "2.2.16",
+    "electron-log": "3.0.8",
     "electron-notarize": "0.1.1",
     "electron-rebuild": "1.8.6",
     "electron-settings": "3.1.4",
@@ -73,8 +73,8 @@
     "ts-node": "8.4.1",
     "tslint": "5.18.0",
     "typescript": "3.5.3",
-    "webpack": "4.29.5",
-    "webpack-cli": "3.2.3"
+    "webpack": "4.41.2",
+    "webpack-cli": "3.3.10"
   },
   "pre-commit": [
     "precommit-msg"
