@@ -1,4 +1,4 @@
-export * from './auto-update-settings';
+export * from './application-settings';
 export * from './command-line-args';
 export * from './notification';
 export * from './ipc-response';
